requirements.txt
quickfix2
loguru